refactor: 테이블 생성해주는 DatabasePopulatorUtils클래스 jdbc 모듈로 이동
