+++ app/src/main/java/com/techcourse/support/context/ContextLoaderListener.java
@@ -1,7 +1,7 @@
 package com.techcourse.support.context;
 
 import com.techcourse.config.DataSourceConfig;
-import com.techcourse.support.jdbc.init.DatabasePopulatorUtils;
+import nextstep.jdbc.DatabasePopulatorUtils;
 import jakarta.servlet.ServletContextEvent;
 import jakarta.servlet.ServletContextListener;
 import jakarta.servlet.annotation.WebListener;
