+++ jdbc/src/main/java/nextstep/jdbc/DatabasePopulatorUtils.java
@@ -1,4 +1,4 @@
-package com.techcourse.support.jdbc.init;
+package nextstep.jdbc;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
